+++ NEWS.md
@@ -18,6 +18,8 @@
 
 ## Internal changes
 
++ Switched internal cat_green_tick to cli::cli_alert_success()
+
 # golem 0.3.1 (CRAN Version)
 
 ## New functions